Fix minor defects

+++ src/util/Util.cpp
@@ -106,7 +106,7 @@ uint64_t fastTickCount()
 #if defined(CLOCK_MONOTONIC_COARSE)
     timespec ts;
     clock_gettime(CLOCK_MONOTONIC_COARSE, &ts);
-    return ts.tv_sec * 1000UL + ts.tv_nsec / 1000000UL;
+    return (uint64_t)ts.tv_sec * 1000UL + ts.tv_nsec / 1000000UL;
 #else
     return tickCount();
 #endif
